// todo-app-frontend/src/api/models/BoardInvitationCreate.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Todo App API
 * API for managing users and their todos. Now with WebSockets!
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { Role } from './Role';
import {
    RoleFromJSON,
    RoleFromJSONTyped,
    RoleToJSON,
    RoleToJSONTyped,
} from './Role';

/**
 * 
 * @export
 * @interface BoardInvitationCreate
 */
export interface BoardInvitationCreate {
    /**
     * 
     * @type {string}
     * @memberof BoardInvitationCreate
     */
    invitedUserEmail: string;
    /**
     * 
     * @type {Role}
     * @memberof BoardInvitationCreate
     */
    role?: Role;
}



/**
 * Check if a given object implements the BoardInvitationCreate interface.
 */
export function instanceOfBoardInvitationCreate(value: object): value is BoardInvitationCreate {
    if (!('invitedUserEmail' in value) || value['invitedUserEmail'] === undefined) return false;
    return true;
}

export function BoardInvitationCreateFromJSON(json: any): BoardInvitationCreate {
    return BoardInvitationCreateFromJSONTyped(json, false);
}

export function BoardInvitationCreateFromJSONTyped(json: any, ignoreDiscriminator: boolean): BoardInvitationCreate {
    if (json == null) {
        return json;
    }
    return {
        
        'invitedUserEmail': json['invited_user_email'],
        'role': json['role'] == null ? undefined : RoleFromJSON(json['role']),
    };
}

export function BoardInvitationCreateToJSON(json: any): BoardInvitationCreate {
    return BoardInvitationCreateToJSONTyped(json, false);
}

export function BoardInvitationCreateToJSONTyped(value?: BoardInvitationCreate | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'invited_user_email': value['invitedUserEmail'],
        'role': RoleToJSON(value['role']),
    };
}